renamed Suggetion baseclass to SuggestionBase
