+++ src/models/SuggestionBase.js
