+++ src/models/SuggestionDaterange.js
@@ -1,4 +1,4 @@
-import Suggestion from '@/models/Suggestion';
+import Suggestion from '@/models/SuggestionBase';
 import Daterange from '@/models/Daterange';
 
 export default class SuggestionDaterange extends Suggestion {
